requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pyautogui>=0.9.50
pillow>=8.0.0